EXEC module init commit

+++ nicnbk-data/nicnbk-data-repo-api/src/main/java/kz/nicnbk/repo/api/lookup/CorpMeetingTypeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 /**
  * Created by zhambyl on 27-Mar-17.
  */
-public interface CorpMeetingTypeRepository extends PagingAndSortingRepository<CorpMeetingType, Long> {
+public interface CorpMeetingTypeRepository extends PagingAndSortingRepository<CorpMeetingType, Integer> {
 
     CorpMeetingType findByCode(String code);
